+++ webextension/js/newTab/newTab-init.js
@@ -1,4 +1,5 @@
 import {theme_update} from "../classes/backgroundTheme.js";
+import {loadTranslations} from '../translation-api.js';
 
 async function init() {
 	import('./newTab.js')
@@ -6,6 +7,8 @@ async function init() {
 	;
 	document.querySelector('#newTab-script')?.remove();
 	theme_update()
+		.catch(console.error);
+	loadTranslations()
 		.catch(console.error);
 }
 init()