+++ webextension/js/newTab/newTab-reopenTab.js
@@ -4,30 +4,60 @@
  */
 async function getLastClosed() {
 	return await chrome.sessions.getRecentlyClosed({
-		maxResults: 1
+		maxResults: 16
 	});
 }
 
-async function reopenLastClosed() {
-	/**
-	 * @type {chrome.sessions.Session|undefined}
-	 */
-	const lastClosedSession = (await getLastClosed()
-		.catch(console.error) ?? []).at(0)
-	;
-	if (lastClosedSession && typeof lastClosedSession.tab?.sessionId === 'string') {
-		await chrome.sessions.restore(lastClosedSession.tab.sessionId);
-	}
+/**
+ *
+ * @param {string} sessionId
+ * @return {Promise<void>}
+ */
+async function reopenLastClosed(sessionId) {
+	await chrome.sessions.restore(sessionId);
 }
 
+/**
+ *
+ * @type {ReturnType<typeof setTimeout>|null}
+ */
+let reopenTabStateRefresh_timer = null;
 export async function reopenTabStateRefresh() {
-	const $reopenWindow = document.querySelector('#reopenWindow');
-	if ($reopenWindow) {
-		const hasLastClosed = ((await getLastClosed().catch(console.error)) ?? []).length > 0;
-		$reopenWindow.disabled = !hasLastClosed;
-	} else {
-		console.error('Button reopenWindow not found!')
+	if (reopenTabStateRefresh_timer) {
+		clearTimeout(reopenTabStateRefresh_timer);
+		reopenTabStateRefresh_timer = null;
 	}
+	reopenTabStateRefresh_timer = setTimeout(_reopenTabStateRefresh, 250);
+}
+async function _reopenTabStateRefresh() {
+	/**
+	 *
+	 * @type {HTMLSelectElement}
+	 */
+	const $reopenWindow = document.querySelector('select#reopenWindow');
+	if (!$reopenWindow) {
+		console.error('Select reopenWindow not found!');
+		return;
+	}
+
+	const lastClosedSessions = (await getLastClosed()
+		.catch(console.error) ?? []);
+
+	for (let option of $reopenWindow.options) {
+		if (option.value) option.remove();
+	}
+
+	for (let lastClosedSession of lastClosedSessions) {
+		$reopenWindow.options.add(
+			lastClosedSession.tab ?
+			new Option(lastClosedSession.tab.title, lastClosedSession.tab.sessionId)
+			:
+			new Option(lastClosedSession.window.tabs.at(0).title ?? lastClosedSession.window.sessionId, lastClosedSession.window.sessionId)
+		);
+	}
+
+	$reopenWindow.disabled = $reopenWindow.options.length <= 1;
+	$reopenWindow.value = '';
 }
 
 chrome.sessions.onChanged.addListener(function onSessionChange() {
@@ -36,11 +66,25 @@ chrome.sessions.onChanged.addListener(function onSessionChange() {
 	;
 });
 
-document.addEventListener('click', function onReopenWindowClick(ev) {
-	const el = ev.target.closest('#reopenWindow');
+document.addEventListener('change', function onReopenWindowChange(ev) {
+	/**
+	 * @type {HTMLSelectElement}
+	 */
+	const el = ev.target.closest('select#reopenWindow');
 	if (!el) return;
 
-	reopenLastClosed()
+	const selectedItem = el.options.item(el.selectedIndex);
+	if (!selectedItem.value) {
+		console.warn('Empty reopenWindow value');
+		return;
+	}
+
+	chrome.sessions.restore(el.value)
 		.catch(console.error)
-	;
+		.finally(() => {
+			selectedItem.remove();
+			el.value = '';
+			reopenTabStateRefresh()
+				.catch(console.error);
+		});
 });