+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# WIP
+* i : Theme rework
+* i : Now use native [util.styleText](https://nodejs.org/api/util.html#utilstyletextformat-text-options) instead of `chalk`
+* \+ : Replace the button to re-open last session by a `select`
+
 # 3.5.0
 * i : Minimum version Firefox 147
 * i : Upgrade libraries